+++ app/src/main/res/layout/activity_external_pebble_js.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<WebView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/configureWebview"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent" />
+<FrameLayout android:id="@+id/webview_placeholder"
+    android:layout_height="match_parent"
+    android:layout_width="match_parent"
+    xmlns:android="http://schemas.android.com/apk/res/android" />
